requirements.txt
streamlit==1.38.0
mysql-connector-python==8.0.33
pandas==2.2.2
numpy==1.26.4
